# Conflicts:
#	app/src/main/java/com/nowfloats/AccrossVerticals/domain/DomainEmailActivity.java
#	appService/src/main/java/com/appservice/model/serviceProduct/CatalogProduct.kt
#	appService/src/main/java/com/appservice/rest/EndPoints.kt
#	appService/src/main/java/com/appservice/rest/TaskCode.kt
#	appService/src/main/java/com/appservice/rest/services/WithFloatTwoRemoteData.kt
#	appService/src/main/java/com/appservice/ui/catalog/catalogProduct/addProduct/ProductDetailFragment.kt
#	appService/src/main/java/com/appservice/ui/catalog/catalogProduct/addProduct/information/ProductInformationFragment.kt
#	appService/src/main/java/com/appservice/ui/catalog/catalogService/addService/information/ServiceInformationFragment.kt
#	appService/src/main/res/values/colors.xml
#	appService/src/main/res/values/strings.xml
#	dashboard/src/main/java/com/dashboard/controller/DashboardActivity.kt
#	dashboard/src/main/java/com/dashboard/model/live/websiteItem/WebsiteActionItem.kt
#	dashboard/src/main/java/com/dashboard/utils/DeepLinkUtil.kt
#	dashboard/src/main/res/layout/item_website_item.xml
#	dashboard/src/main/res/layout/website_item_feature.xml
#	florisboard/src/main/java/dev/patrickgold/florisboard/customization/BusinessFeaturesManager.kt
#	florisboard/src/main/java/dev/patrickgold/florisboard/customization/viewholder/FloatUpdateViewHolder.kt
#	framework/src/main/java/com/framework/utils/Util.kt
#	framework/src/main/java/com/framework/views/bottombar/NiceBottomBar.kt

+++ inventoryorder/src/main/java/com/inventoryorder/model/orderRequest/OrderInitiateRequestNew.kt
@@ -74,7 +74,7 @@ data class ItemsItemNew(
 	var imageUri: String? = null,
 
 	@field:SerializedName("DiscountAmount")
-	var discountAmount: Int? = null,
+	var discountAmount: Double? = null,
 
 	@field:SerializedName("Name")
 	var name: String? = null,
@@ -86,7 +86,7 @@ data class ItemsItemNew(
 	var type: String? = null,
 
 	@field:SerializedName("Price")
-	var price: Int? = null,
+	var price: Double? = null,
 
 	@field:SerializedName("ProductDetails")
 	var productDetails: ProductDetailsN? = null,
@@ -95,7 +95,7 @@ data class ItemsItemNew(
 	var isAvailable: Boolean? = null,
 
 	@field:SerializedName("ShippingCost")
-	var shippingCost: Int? = null,
+	var shippingCost: Double? = null,
 
 	@field:SerializedName("_id")
 	var id: String? = null,
@@ -269,7 +269,7 @@ data class ShippingDetailsN(
 	var currencyCode: String? = null,
 
 	@field:SerializedName("ShippingCost")
-	var shippingCost: Int? = null,
+	var shippingCost: Double? = null,
 
 	@field:SerializedName("DeliveryMode")
 	var deliveryMode: String? = null,
